+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - [`TranscludeToken.prototype.getFrame`](https://github.com/bhsd-harry/wikiparser-node/wiki/TranscludeToken-%28EN%29#getframe) method that returns a basic JSON [frame object](https://www.mediawiki.org/wiki/Extension:Scribunto/Lua_reference_manual#frame-object) for the `#invoke` parser function
 - [`Parser.setFunctionHook`](https://github.com/bhsd-harry/wikiparser-node/wiki/Parser-%28EN%29#setfunctionhook) method that defines a custom parser function hook for the [`Token.prototype.expand`](https://github.com/bhsd-harry/wikiparser-node/wiki/Token-%28EN%29#expand) method
 - [`Parser.setHook`](https://github.com/bhsd-harry/wikiparser-node/wiki/Parser-%28EN%29#sethook) method that defines a custom extension tag hook for the [`Token.prototype.expand`](https://github.com/bhsd-harry/wikiparser-node/wiki/Token-%28EN%29#expand) method
+- [`Title.displayTitle`](https://github.com/bhsd-harry/wikiparser-node/wiki/Title-%28EN%29#displaytitle)
 
 **Fixed**